requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0.post1
motor==3.3.1
python-dotenv==1.0.0
openai==0.28.1
PyJWT==2.8.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
pydantic==2.5.2
pydantic-settings==2.1.0
bcrypt==4.0.1
